requirements.txt
PyQt6 == 6.2.3
matplotlib==3.5.1
Pillow==9.0.1
numpy==1.21.5
svglib==1.2.1
docopt==0.6.2
# svglib==1.2.1
svgwrite==1.4.1
imageio==2.16.1